Pass XLets as input parameters for lineage runner
