+++ ingestion/src/airflow_provider_openmetadata/lineage/backend.py
@@ -23,6 +23,7 @@ from airflow_provider_openmetadata.lineage.config.loader import (
     get_lineage_config,
 )
 from airflow_provider_openmetadata.lineage.runner import AirflowLineageRunner
+from airflow_provider_openmetadata.lineage.xlets import XLets, get_xlets_from_dag
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 
 
@@ -69,12 +70,13 @@ class OpenMetadataLineageBackend(LineageBackend):
         try:
             config: AirflowLineageConfig = get_lineage_config()
             metadata = OpenMetadata(config.metadata_config)
+            xlets: XLets = get_xlets_from_dag(context["dag"])
 
             runner = AirflowLineageRunner(
                 metadata=metadata,
                 service_name=config.airflow_service_name,
                 dag=context["dag"],
-                context=context,
+                xlets=xlets,
                 only_keep_dag_lineage=config.only_keep_dag_lineage,
                 max_status=config.max_status,
             )
